refactor structure task module to include ngrx logic there
