+++ src/app/task/state/effects/task.effects.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@angular/core';
-import { TaskService } from './../../task/services/task.service';
+import { TaskService } from './../../services/task.service';
 import { Effect, Actions } from '@ngrx/effects';
 import * as taskActions from './../actions/task.actions';
 import { switchMap, map, tap } from 'rxjs/operators';
-import { Task } from './../../task/model/task';
+import { Task } from './../../model/task';
 
 @Injectable()
 export class TaskEffects {
