+++ src/app/task/state/models/app-state.ts
@@ -1,4 +1,4 @@
-import { Task } from './../../task/model/task';
+import { Task } from './../../model/task';
 
 export interface IAppState {
   tasks: Task[];
